+++ src/components/box/box.d.ts
@@ -1,0 +1,9 @@
+import * as React from 'react';
+import { SpacingProps, ColorProps, LayoutProps, FlexBoxProps } from '../../utils/helpers/options-helper';
+
+export interface BoxProps extends SpacingProps, ColorProps, LayoutProps, FlexBoxProps {
+    as?: React.ElementType;
+}
+
+declare const Box: React.FunctionComponent<BoxProps>;
+export default Box;